fixed bug with threshold and improved tests
I found that I didn't have threshold applied correctly, this is now correct. I also fixed the multiplicative constant so that the DFT matches what is expected from analytical (there is a 1/sqrt(2pi), expected in common definitions of FT including that in the test). Threshold applied in est_spec_norm and Fourier transform is corrected in est_spec_norm_from_data.